Move the setups folder from the tests folder
