+++ src/setups/setupTests.js
@@ -1,4 +1,4 @@
 import Enzyme from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
+import Adapter from 'enzyme-adapter-react-17-updated';
 
 Enzyme.configure({ adapter: new Adapter() });
